fix(build): remove warning

+++ OpenGLNgine/OpenGLNgine/Render/RenderWindow.cpp
@@ -13,9 +13,9 @@ void RenderWindow::render() const
     for(const auto& vp : m_viewports)
     {
         const auto& size = vp.second->getViewport();
-        ::GL::Rasterizer::setViewport(size[0]*static_cast<float>(m_width), size[1]*static_cast<float>(m_height), size[2]*static_cast<float>(m_width), size[3]*static_cast<float>(m_height));
+        ::GL::Rasterizer::setViewport(static_cast< int >(size[0]*m_width), static_cast< int >(size[1]*m_height), static_cast< int >(size[2]*m_width), static_cast< int >(size[3]*m_height));
         ::GL::Rasterizer::enableScissorTest(true);
-        ::GL::Rasterizer::setScissor(size[0]*static_cast<float>(m_width), size[1]*static_cast<float>(m_height), size[2]*static_cast<float>(m_width), size[3]*static_cast<float>(m_height));
+        ::GL::Rasterizer::setScissor(static_cast< int >(size[0]*m_width), static_cast< int >(size[1]*m_height), static_cast< int >(size[2]*m_width), static_cast< int >(size[3]*m_height));
         const auto& color = vp.second->getClearColor();
 
         ::GL::PixelOperation::setColorClearValue(color[0], color[1], color[2], color[3]);
